+++ src/dgisim/card/cards_set.py
@@ -20,6 +20,7 @@ _DEFAULT_CARDS: list[type[Card]] = [
     EmbraceOfWinds,
     FloralSidewinder,
     GalesOfReverie,
+    GloriousSeason,
     GrandExpectation,
     IGotYourBack,
     KantenSenmyouBlessing,